Moved helper function, moved node.onInput inside helper
