+++ telos-push/telos-push.html
@@ -1,8 +1,6 @@
 
 <script type="text/javascript">
 
-    const helper = require('../lib/helper.js'); // Import functions
-
     // Set nodeid which will also be our eosio primary index (name converts to uint64_t)
     var nodeid = '';
     // subset of characters allowed in eosio names
